+++ src/app/(authorized)/Notifications/components/DiscardNotificationBtn.js
