+++ src/app/(authorized)/games/[game_name]/[platform]/[ign]/components/Apex/RenderedLegends.js
