+++ src/app/(authorized)/games/[game_name]/[platform]/[ign]/components/Division-2/OtherSections.js
