+++ src/app/(authorized)/games/[game_name]/[platform]/[ign]/components/Splitgate/SplitgateTabs.js
