+++ src/app/(authorized)/games/[game_name]/[platform]/[ign]/page.js
@@ -1,4 +1,4 @@
-import RefreshStatsBtn from "@/components/GameStats/RefreshStatsBtn";
+import RefreshStatsBtn from "@/app/(authorized)/games/[game_name]/[platform]/[ign]/components/RefreshStatsBtn";
 import { CurrentlySupportedGames } from "@/app/(authorized)/SupportedGames/page";
 import { redirect } from "next/navigation";
 import { TRNProfile } from "@/app/utils/external-apis/externalApi";
@@ -6,9 +6,9 @@ import Image from "next/image";
 import { SiOrigin, SiSteam, SiPlaystation, SiUbisoft } from "react-icons/si";
 import { FaXbox } from "react-icons/fa";
 import { BsEyeFill } from "react-icons/bs";
-import GameBackgroundImage from "@/components/GameStats/GameBackgroundImage";
-import StatTabs from "@/components/GameStats/StatTabs";
-import IgnNotFound from "@/components/GameStats/IgnNotFound";
+import GameBackgroundImage from "@/app/(authorized)/games/[game_name]/[platform]/[ign]/components/GameBackgroundImage";
+import StatTabs from "@/app/(authorized)/games/[game_name]/[platform]/[ign]/components/StatTabs";
+import IgnNotFound from "@/app/(authorized)/games/[game_name]/[platform]/[ign]/components/IgnNotFound";
 import getGameProps from "@/lib/getGameProps";
