+++ src/components/Navbar/Navbar.js
@@ -5,7 +5,6 @@ import { Input } from "@/components/ui/input"
 import { Search } from 'lucide-react';
 
 
-
 const Navbar = () => {
 
     return(
